mcp-server/pyproject.toml
[project]
name = "mcp-server"
version = "0.1.0"
description = "Shared MCP tool server deployed as a Databricks App"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn>=0.29.0",
    "pydantic>=2.6.0",
    "fastmcp",
    "mcp>=1.26.0",
    "databricks-mcp",
    "databricks-sdk>=0.40.0",
]

[project.scripts]
start-server = "server.main:main"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
